--- src/main/java/kitchenpos/dao/OrderTableRepository.java
@@ -1,25 +1,0 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.OrderTable;
-import org.springframework.dao.EmptyResultDataAccessException;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-
-import javax.sql.DataSource;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-
-public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
-
-    List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
-
-    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
-}
