+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -1,0 +1,11 @@
+package kitchenpos.repository;
+
+import kitchenpos.domain.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+    int countByIdIn(List<Long> menuIds);
+}
