+++ src/main/java/kitchenpos/repository/OrderTableRepository.java
@@ -1,0 +1,13 @@
+package kitchenpos.repository;
+
+import kitchenpos.domain.OrderTable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
+
+    List<OrderTable> findAllByIdIn(List<Long> orderTableIds);
+
+    List<OrderTable> findAllByTableGroupId(Long tableGroupId);
+}
