+++ src/main/java/kitchenpos/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.repository;
+
+import kitchenpos.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
