+++ src/main/java/kitchenpos/repository/TableGroupRepository.java
@@ -1,0 +1,8 @@
+package kitchenpos.repository;
+
+import kitchenpos.domain.TableGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+
+}
