+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.dao.MenuRepository;
+import kitchenpos.repository.MenuRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
